third_party/WebKit/LayoutTests/http/tests/devtools/elements/event-listeners-about-blank.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/elements-test.js"></script>
<script>

function setupEventListeners()
{
    function f() {}
    var frame = document.getElementById("myframe");
    var body = frame.contentDocument.body;
    body.addEventListener("click", f, true);
    var div = frame.contentDocument.createElement("div");
    div.id = "div-in-iframe";
    div.addEventListener("hover", f, {capture: true, once: true});
    body.appendChild(div);
    body.addEventListener("wheel", f, {"passive": true});
}

function test() {
  Common.settingForTest('showEventListenersForAncestors').set(true);
  TestRunner.evaluateInPage('setupEventListeners()', step1);

  function step1() {
    ElementsTestRunner.selectNodeWithId('div-in-iframe', step2);
  }

  function step2() {
    ElementsTestRunner.expandAndDumpSelectedElementEventListeners(TestRunner.completeTest.bind(InspectorTest));
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests event listeners output in the Elements sidebar panel when the listeners are added on an element in about:blank page.
</p>

<iframe id="myframe"></iframe>

</body>
</html>
