third_party/WebKit/LayoutTests/http/tests/devtools/elements/event-listeners-about-blank-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests event listeners output in the Elements sidebar panel when the listeners are added on an element in about:blank page.


======== click ========
== Raw
[expanded] bodyRemoveevent-listeners-about-blank.js:16
    useCapture: true
    passive: false
    once: false
    handler: function f() {}

======== hover ========
== Raw
[expanded] div#div-in-iframeRemoveevent-listeners-about-blank.js:16
    useCapture: true
    passive: false
    once: true
    handler: function f() {}

======== wheel ========
== Raw
[expanded] bodyRemoveToggle Passiveevent-listeners-about-blank.js:16
    useCapture: false
    passive: true
    once: false
    handler: function f() {}
